+++ src/main/java/com/officePharmaceutiqueNationale/OPN/sercice/impl/SpecialitePharmaceutiqueServiceImpl.java
@@ -1,40 +1,47 @@
 package com.officePharmaceutiqueNationale.OPN.sercice.impl;
 
 import com.officePharmaceutiqueNationale.OPN.dto.SpecialitePharmaceutiqueDto;
+import com.officePharmaceutiqueNationale.OPN.exception.ResourceNotFoundException;
+import com.officePharmaceutiqueNationale.OPN.mapper.DciMapper;
 import com.officePharmaceutiqueNationale.OPN.mapper.SpecialitePharmaceutiqueMapper;
 import com.officePharmaceutiqueNationale.OPN.model.Dci;
 import com.officePharmaceutiqueNationale.OPN.model.SpecialitePharmaceutique;
 import com.officePharmaceutiqueNationale.OPN.repository.DciRepository;
 import com.officePharmaceutiqueNationale.OPN.repository.SpecialitePharmaceutiqueRepository;
 import com.officePharmaceutiqueNationale.OPN.sercice.SpecialitePharmaceutiqueService;
+import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
+@Transactional
 public class SpecialitePharmaceutiqueServiceImpl implements SpecialitePharmaceutiqueService {
 
     private final SpecialitePharmaceutiqueRepository specialiteRepository;
-    private final DciRepository dciRepository;
     private final SpecialitePharmaceutiqueMapper specialiteMapper;
+    private final DciRepository dciRepository;
 
-    public SpecialitePharmaceutiqueServiceImpl(SpecialitePharmaceutiqueRepository specialiteRepository, DciRepository dciRepository, SpecialitePharmaceutiqueMapper specialiteMapper) {
+    public SpecialitePharmaceutiqueServiceImpl(SpecialitePharmaceutiqueRepository specialiteRepository,
+                                               SpecialitePharmaceutiqueMapper specialiteMapper, DciRepository dciRepository, DciMapper dciMapper) {
         this.specialiteRepository = specialiteRepository;
-        this.dciRepository = dciRepository;
         this.specialiteMapper = specialiteMapper;
+        this.dciRepository = dciRepository;
     }
 
+    // Enregistrer une spécialité pharmaceutique
     @Override
-    public SpecialitePharmaceutiqueDto enregistrerUneSpecialite(SpecialitePharmaceutiqueDto specialiteDto, String idDci) {
+    public SpecialitePharmaceutiqueDto enregistrerUneSpecialite(@Valid SpecialitePharmaceutiqueDto specialiteDto) {
 
-        // Recherche du DCI associé
-        Dci dci = dciRepository.findById(idDci)
-                .orElseThrow(() -> new RuntimeException("DCI non trouvé avec l'ID : " + idDci));
+        if (specialiteDto == null || specialiteDto.getLibelleSpecialiteMedicament() == null
+                || specialiteDto.getLibelleSpecialiteMedicament().isBlank()) {
+            throw new IllegalArgumentException("Le DTO de la spécialité pharmaceutique ne doit pas être nul et doit contenir un libellé valide");
+        }
 
         // Génération du UUID pour l'ID de la spécialité pharmaceutique
         String specialiteId = UUID.randomUUID().toString();
@@ -42,7 +49,23 @@ public class SpecialitePharmaceutiqueServiceImpl implements SpecialitePharmaceut
         // Création de l'entité SpecialitePharmaceutique avec l'ID généré
         SpecialitePharmaceutique specialite = specialiteMapper.toEntity(specialiteDto);
         specialite.setId(specialiteId);
-        specialite.setDci(dci);
+
+        // Attribut isDeleted défini à True
+        specialite.setIsDeleted(true);
+
+        // Vérifier si le DCI est présent
+        if (specialiteDto.getDci() != null && specialiteDto.getDci().getId() != null) {
+            Optional<Dci> optionalDci = dciRepository.findById(specialiteDto.getDci().getId());
+            Dci dci = optionalDci.orElseThrow(() ->
+                    new IllegalArgumentException("DCI avec l'ID " + specialiteDto.getDci().getId() + " n'existe pas.")
+            );
+
+            // Mettre à jour l'attribut isDeleted du DCI en False empêchant sa suppression
+            dci.setIsDeleted(false);
+            dciRepository.save(dci);
+
+            specialite.setDci(dci);
+        }
 
         // Enregistrer la spécialité dans la base de données
         SpecialitePharmaceutique savedSpecialite = specialiteRepository.save(specialite);
@@ -51,55 +74,96 @@ public class SpecialitePharmaceutiqueServiceImpl implements SpecialitePharmaceut
         return specialiteMapper.toDto(savedSpecialite);
     }
 
-    @Override
-    public SpecialitePharmaceutiqueDto modifierUneSpecialite(SpecialitePharmaceutiqueDto specialiteDto) {
-        SpecialitePharmaceutique specialite = specialiteRepository.findById(specialiteDto.getId())
-                .orElseThrow(() -> new RuntimeException("Spécialité non trouvée avec l'ID : " + specialiteDto.getId()));
 
+    // Modifier une spécialité existante
+    @Override
+    public SpecialitePharmaceutiqueDto modifierUneSpecialite(@Valid SpecialitePharmaceutiqueDto specialiteDto) {
+
+        if (specialiteDto == null || specialiteDto.getId() == null || specialiteDto.getId().isBlank()) {
+            throw new IllegalArgumentException("Le DTO de la spécialité pharmaceutique ne doit pas être nul et doit contenir un ID valide");
+        }
+
+        // Récupérer la spécialité existante
+        SpecialitePharmaceutique specialite = specialiteRepository.findById(specialiteDto.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Spécialité non trouvée avec l'ID : " + specialiteDto.getId()));
+
+        // Modification des attributs
         specialite.setLibelleSpecialiteMedicament(specialiteDto.getLibelleSpecialiteMedicament());
         specialite.setNomLaboratoire(specialiteDto.getNomLaboratoire());
 
+        // Modification du DCI si nécessaire
+        if (specialiteDto.getDci() != null && specialiteDto.getDci().getId() != null) {
+
+            // Vérifier si le DCI existe
+            Optional<Dci> optionalDci = dciRepository.findById(specialiteDto.getDci().getId());
+            Dci dci = optionalDci.orElseThrow(() ->
+                    new IllegalArgumentException("DCI avec l'ID " + specialiteDto.getDci().getId() + " n'existe pas.")
+            );
+
+            // Mettre à jour le DCI de la spécialité
+            specialite.setDci(dci);
+        }
+
+        // Enregistrer les modifications
         SpecialitePharmaceutique updatedSpecialite = specialiteRepository.save(specialite);
+
         return specialiteMapper.toDto(updatedSpecialite);
     }
 
+
+    // Supprimer une spécialité pharmaceutique
     @Override
     public void supprimerUneSpecialite(String id) {
+        if (id == null || id.isBlank()) {
+            throw new IllegalArgumentException("L'ID de la spécialité pharmaceutique ne doit pas être nul ou vide");
+        }
 
         SpecialitePharmaceutique specialite = specialiteRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Spécialité non trouvée avec l'ID : " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Spécialité non trouvée avec l'ID : " + id));
+
+        // Vérifier l'état de l'attribut isDeleted
+        if (Boolean.FALSE.equals(specialite.getIsDeleted())) {
+            throw new IllegalStateException("La spécialité ne peut être supprimée que si son état est marqué comme supprimé (isDeleted = true)");
+        }
 
         specialiteRepository.delete(specialite);
     }
 
+    // Récupérer toutes les spécialités pharmaceutiques
     @Override
     public List<SpecialitePharmaceutiqueDto> recupererLesSpecialites() {
         List<SpecialitePharmaceutique> specialites = specialiteRepository.findAll();
-        return specialites.stream()
-                .map(specialiteMapper::toDto)
-                .collect(Collectors.toList());
+        return specialiteMapper.toDtoList(specialites);
     }
 
+    // Récupérer une spécialité par id
     @Override
     public SpecialitePharmaceutiqueDto recupererUneSpecialiteParId(String id) {
+        if (id == null || id.isBlank()) {
+            throw new IllegalArgumentException("L'ID de la spécialité pharmaceutique ne peut pas être nul ou vide");
+        }
+
         SpecialitePharmaceutique specialite = specialiteRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Spécialité non trouvée avec l'ID : " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Spécialité non trouvée avec l'ID : " + id));
 
         return specialiteMapper.toDto(specialite);
     }
 
+    // Faire la pagination
     @Override
     public Page<SpecialitePharmaceutiqueDto> recuperationParPagination(int page, int limit) {
-        Pageable pageable = PageRequest.of(page, limit);
-        Page<SpecialitePharmaceutique> specialitePage = specialiteRepository.findAll(pageable);
+        if (page < 0 || limit <= 0) {
+            throw new IllegalArgumentException("Les paramètres de pagination ne sont pas valides");
+        }
+
+        Page<SpecialitePharmaceutique> specialitePage = specialiteRepository.findAll(PageRequest.of(page, limit));
         return specialitePage.map(specialiteMapper::toDto);
     }
 
+    // Récupérer les métadonnées
     @Override
     public Page<SpecialitePharmaceutiqueDto> recuperationDesMetadonnees(int page, int limit) {
-        Pageable pageable = PageRequest.of(page, limit);
-        Page<SpecialitePharmaceutique> specialitePage = specialiteRepository.findAll(pageable);
-        return specialitePage.map(specialiteMapper::toDto);
+        // Implémentation spécifique selon vos besoins
+        throw new UnsupportedOperationException("Méthode recuperationDesMetadonnees non implémentée");
     }
 }
-
